+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 # 0.13.0-alpha.8
 
 * Fix CLI `table migrate-transform-tables` for complex case
+* Add magic `ds` inject into `BatchGenerate`
 
 # 0.13.0-alpha.7